+++ cocoa-beans-spigot/src/main/java/net/apartium/cocoabeans/spigot/visibility/VisibilityListener.java
@@ -19,7 +19,6 @@ import org.bukkit.event.player.PlayerQuitEvent;
 
     @EventHandler(priority = EventPriority.NORMAL)
     public void onPlayerJoin(PlayerJoinEvent event) {
-        manager.getPlayer(event.getPlayer()).clearPlayerRef();
         manager.handlePlayerJoin(event.getPlayer());
     }